+++ src/main/java/com/aws/greengrass/telemetry/nucleus/emitter/NucleusEmitterConfiguration.java
@@ -7,17 +7,35 @@ package com.aws.greengrass.telemetry.nucleus.emitter;
 
 import com.aws.greengrass.logging.api.Logger;
 import com.aws.greengrass.util.Coerce;
+import com.aws.greengrass.util.Utils;
 import lombok.Builder;
 import lombok.Value;
 import org.apache.commons.lang3.BooleanUtils;
 
+import java.util.Collections;
+import java.util.List;
+import java.util.Locale;
 import java.util.Map;
 
 import static com.aws.greengrass.telemetry.nucleus.emitter.Constants.CONFIG_INVALID_OPTION_ERROR_LOG;
+import static com.aws.greengrass.telemetry.nucleus.emitter.Constants.DEFAULT_METRICS_LEVEL;
+import static com.aws.greengrass.telemetry.nucleus.emitter.Constants.DEFAULT_OUTPUT_DIRECTORY;
+import static com.aws.greengrass.telemetry.nucleus.emitter.Constants.DEFAULT_OUTPUT_MODE;
 import static com.aws.greengrass.telemetry.nucleus.emitter.Constants.DEFAULT_TELEMETRY_PUBLISH_INTERVAL_MS;
 import static com.aws.greengrass.telemetry.nucleus.emitter.Constants.DEFAULT_TELEMETRY_PUBSUB_TOPIC;
+import static com.aws.greengrass.telemetry.nucleus.emitter.Constants.DETAILED_METRICS_LEVEL;
+import static com.aws.greengrass.telemetry.nucleus.emitter.Constants.EXCLUDE_INTERFACES_CONFIG_NAME;
+import static com.aws.greengrass.telemetry.nucleus.emitter.Constants.EXCLUDE_MOUNTS_CONFIG_NAME;
+import static com.aws.greengrass.telemetry.nucleus.emitter.Constants.METRICS_LEVEL_CONFIG_NAME;
+import static com.aws.greengrass.telemetry.nucleus.emitter.Constants.METRICS_LEVEL_CONFIG_PARSE_ERROR_LOG;
 import static com.aws.greengrass.telemetry.nucleus.emitter.Constants.MQTT_TOPIC_CONFIG_NAME;
 import static com.aws.greengrass.telemetry.nucleus.emitter.Constants.MQTT_TOPIC_CONFIG_PARSE_ERROR_LOG;
+import static com.aws.greengrass.telemetry.nucleus.emitter.Constants.OUTPUT_DIRECTORY_CONFIG_NAME;
+import static com.aws.greengrass.telemetry.nucleus.emitter.Constants.OUTPUT_MODE_BOTH;
+import static com.aws.greengrass.telemetry.nucleus.emitter.Constants.OUTPUT_MODE_CONFIG_NAME;
+import static com.aws.greengrass.telemetry.nucleus.emitter.Constants.OUTPUT_MODE_CONFIG_PARSE_ERROR_LOG;
+import static com.aws.greengrass.telemetry.nucleus.emitter.Constants.OUTPUT_MODE_EMF;
+import static com.aws.greengrass.telemetry.nucleus.emitter.Constants.OUTPUT_MODE_IPC;
 import static com.aws.greengrass.telemetry.nucleus.emitter.Constants.PUBSUB_PUBLISH_CONFIG_NAME;
 import static com.aws.greengrass.telemetry.nucleus.emitter.Constants.PUBSUB_PUBLISH_CONFIG_PARSE_ERROR_LOG;
 import static com.aws.greengrass.telemetry.nucleus.emitter.Constants.PUBSUB_TOPIC_CONFIG_NAME;
@@ -26,7 +44,7 @@ import static com.aws.greengrass.telemetry.nucleus.emitter.Constants.TELEMETRY_P
 import static com.aws.greengrass.telemetry.nucleus.emitter.Constants.TELEMETRY_PUBLISH_INTERVAL_CONFIG_PARSE_ERROR_LOG;
 
 @Value
-@Builder
+@Builder(toBuilder = true)
 public class NucleusEmitterConfiguration {
 
     //Configurable options
@@ -40,6 +58,25 @@ public class NucleusEmitterConfiguration {
     String mqttTopic = "";
     @Builder.Default
     long telemetryPublishIntervalMs = DEFAULT_TELEMETRY_PUBLISH_INTERVAL_MS;
+
+    @Builder.Default
+    String metricsLevel = DEFAULT_METRICS_LEVEL;
+    @Builder.Default
+    String outputMode = DEFAULT_OUTPUT_MODE;
+    @Builder.Default
+    String outputDirectory = DEFAULT_OUTPUT_DIRECTORY;
+    @Builder.Default
+    List<String> excludeMounts = Collections.emptyList();
+    @Builder.Default
+    List<String> excludeInterfaces = Collections.emptyList();
+
+    public boolean isDetailedMetrics() {
+        return DETAILED_METRICS_LEVEL.equals(metricsLevel);
+    }
+
+    public boolean isEmfEnabled() {
+        return OUTPUT_MODE_EMF.equals(outputMode) || OUTPUT_MODE_BOTH.equals(outputMode);
+    }
 
     /**
      * Get the Nucleus Emitter configuration from the POJO map.
@@ -98,6 +135,44 @@ public class NucleusEmitterConfiguration {
                         logger.error(PUBSUB_TOPIC_CONFIG_PARSE_ERROR_LOG, entry.getValue());
                         return null;
                     }
+                case METRICS_LEVEL_CONFIG_NAME:
+                    String mlVal = Coerce.toString(entry.getValue())
+                            .toLowerCase(Locale.ROOT);
+                    if (mlVal.isEmpty()) {
+                        break;
+                    }
+                    if (DEFAULT_METRICS_LEVEL.equals(mlVal)
+                            || DETAILED_METRICS_LEVEL.equals(mlVal)) {
+                        config.metricsLevel(mlVal);
+                        break;
+                    }
+                    logger.error(METRICS_LEVEL_CONFIG_PARSE_ERROR_LOG, entry.getValue());
+                    return null;
+                case OUTPUT_MODE_CONFIG_NAME:
+                    String omVal = Coerce.toString(entry.getValue())
+                            .toLowerCase(Locale.ROOT);
+                    if (omVal.isEmpty()) {
+                        break;
+                    }
+                    if (OUTPUT_MODE_IPC.equals(omVal) || OUTPUT_MODE_EMF.equals(omVal)
+                            || OUTPUT_MODE_BOTH.equals(omVal)) {
+                        config.outputMode(omVal);
+                        break;
+                    }
+                    logger.error(OUTPUT_MODE_CONFIG_PARSE_ERROR_LOG, entry.getValue());
+                    return null;
+                case OUTPUT_DIRECTORY_CONFIG_NAME:
+                    String dirVal = Coerce.toString(entry.getValue());
+                    if (!Utils.isEmpty(dirVal)) {
+                        config.outputDirectory(dirVal);
+                    }
+                    break;
+                case EXCLUDE_MOUNTS_CONFIG_NAME:
+                    config.excludeMounts(Coerce.toStringList(entry.getValue()));
+                    break;
+                case EXCLUDE_INTERFACES_CONFIG_NAME:
+                    config.excludeInterfaces(Coerce.toStringList(entry.getValue()));
+                    break;
                 default:
                     logger.error(CONFIG_INVALID_OPTION_ERROR_LOG, entry.getKey());
                     return null;
@@ -106,4 +181,5 @@ public class NucleusEmitterConfiguration {
 
         return config.build();
     }
+
 }